requirements.txt
amqp==2.5.2
appnope==0.1.0
asgiref==3.2.7
backcall==0.1.0
backoff==1.6.0
billiard==3.6.3.0
celery==4.4.2
celery-redbeat==0.13.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9
decorator==4.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.2.1
django-dotenv==1.4.2
django-extensions==2.2.9
django-loginas==0.3.8
django-slack==5.14.3
django-stubs==1.5.0
djangorestframework==3.11.0
djangorestframework-csv==2.1.0
djangorestframework-stubs==1.1.0
freezegun==0.3.14
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.6.0
inflection==0.3.1
ipdb==0.12.2
ipython==7.9.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.16.0
Jinja2==2.11.1
kombu==4.6.8
MarkupSafe==1.1.1
monotonic==1.5
mypy==0.761
mypy-extensions==0.4.3
numpy==1.18.1
oauthlib==3.1.0
packaging==20.1
pandas==1.0.3
parso==0.6.1
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
posthoganalytics==1.0.10
prompt-toolkit==2.0.10
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.5.2
PyJWT==1.7.1
pyparsing==2.4.6
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
redis==3.4.1
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sentry-sdk==0.14.3
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.0
tblib==1.6.0
tenacity==6.1.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.9
Werkzeug==1.0.0
zipp==3.1.0
